requirements.txt
streamlit
sentence-transformers
faiss-cpu
requests
python-dotenv
numpy
groq
pypdf
speechrecognition
pyttsx3
pandas
streamlit-mic-recorder


